chore: use api modeling for rack firmware API management
This is similar to https://github.com/NVIDIA/ncx-infra-controller-core/pull/602, https://github.com/NVIDIA/ncx-infra-controller-core/pull/598, and https://github.com/NVIDIA/ncx-infra-controller-core/pull/596, in which I'm working through places where we should be implementing our new `STYLE_GUIDE.md` guidelines, and ensuring proper models exist in `api-model` for taking gRPC requests, translating them at the API layer with `From/TryFrom` implementations, and passing them downstream to the `api-db` layer.
