+++ crates/api-model/src/rack_firmware.rs
@@ -17,6 +17,51 @@
 
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
+use sqlx::postgres::PgRow;
+use sqlx::types::Json;
+use sqlx::{FromRow, Row};
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct RackFirmware {
+    pub id: String,
+    pub config: Json<serde_json::Value>,
+    pub available: bool,
+    pub parsed_components: Option<Json<serde_json::Value>>,
+    pub created: DateTime<Utc>,
+    pub updated: DateTime<Utc>,
+}
+
+impl<'r> FromRow<'r, PgRow> for RackFirmware {
+    fn from_row(row: &'r PgRow) -> Result<Self, sqlx::Error> {
+        Ok(RackFirmware {
+            id: row.try_get("id")?,
+            config: row.try_get("config")?,
+            available: row.try_get("available")?,
+            parsed_components: row.try_get("parsed_components")?,
+            created: row.try_get("created")?,
+            updated: row.try_get("updated")?,
+        })
+    }
+}
+
+impl From<&RackFirmware> for rpc::forge::RackFirmware {
+    fn from(db: &RackFirmware) -> Self {
+        let parsed_components = db
+            .parsed_components
+            .as_ref()
+            .map(|p| p.0.to_string())
+            .unwrap_or_else(|| "{}".to_string());
+
+        rpc::forge::RackFirmware {
+            id: db.id.clone(),
+            config_json: db.config.0.to_string(),
+            available: db.available,
+            created: db.created.format("%Y-%m-%d %H:%M:%S").to_string(),
+            updated: db.updated.format("%Y-%m-%d %H:%M:%S").to_string(),
+            parsed_components,
+        }
+    }
+}
 
 /// A record of a rack firmware apply operation
 #[derive(Debug, Clone, Serialize, Deserialize)]
